# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Inventory Management API
  version: 1.0.0
  description: API for managing inventory, products, users, and other related entities.
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
paths:
...
  /bulk/products/template:
    get:
      summary: Download product import template
      description: Downloads a CSV/Excel template file with required headers for product creation
      tags:
        - Bulk Operations
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: CSV template file
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '500':
          $ref: '#/components/responses/InternalServerError'
  /bulk/products/import:
    post:
      summary: Upload a file for bulk product import
      description: Uploads a CSV/Excel file for bulk product creation/update. Returns a job ID for status tracking.
      tags:
        - Bulk Operations
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Bulk import job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkImportJobStatus'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /bulk/products/import/{jobId}/status:
    get:
      summary: Get bulk import job status
      description: Retrieves the status and validation results of a bulk import job
      tags:
        - Bulk Operations
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: jobId
          schema:
            type: string
            format: uuid
          required: true
          description: Bulk Import Job ID
      responses:
        '200':
          description: Bulk import job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkImportJobStatus'
        '404':
          $ref: '#/components/responses/NotFound'
  /bulk/products/import/{jobId}/confirm:
    post:
      summary: Confirm and execute bulk import
      description: Confirms and executes the bulk import after preview
      tags:
        - Bulk Operations
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: jobId
          schema:
            type: string
            format: uuid
          required: true
          description: Bulk Import Job ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkImportConfirmRequest'
      responses:
        '200':
          description: Bulk import confirmation status
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                    format: uuid
                    example: "a1b2c3d4-e5f6-7890-1234-567890abcdef"
                  status:
                    type: string
                    example: PROCESSING
                  message:
                    type: string
                    example: Bulk import initiated
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /bulk/products/export:
    get:
      summary: Export product catalog
      description: Exports the entire product catalog or a filtered list of products to a CSV/Excel file
      tags:
        - Bulk Operations
      security:
        - ApiKeyAuth: []
      parameters:
        - in: query
          name: format
          schema:
            type: string
            enum: [csv, excel]
            default: csv
          description: Export format
        - in: query
          name: category
          schema:
            type: integer
          description: Filter by Category ID
        - in: query
          name: supplier
          schema:
            type: integer
          description: Filter by Supplier ID
      responses:
        '200':
          description: Exported product data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkExportJobStatus'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BulkImportConfirmRequest:
      type: object
      required:
        - jobId
      properties:
        jobId:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-1234-567890abcdef"
    BulkImportJobStatus:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-1234-567890abcdef"
        status:
          type: string
          enum: [QUEUED, PROCESSING, PENDING_CONFIRMATION, COMPLETED, FAILED]
          example: QUEUED
        message:
          type: string
          example: Bulk import job queued for processing.
        filePath:
          type: string
          example: ./uploads/products.csv
        totalRecords:
          type: integer
          example: 100
        validRecords:
          type: integer
          example: 95
        invalidRecords:
          type: integer
          example: 5
        errors:
          type: array
          items:
            type: object
            properties:
              row:
                type: integer
                example: 2
              message:
                type: string
                example: Invalid SKU format
        preview:
          type: array
          items:
            type: object # Example of a product data row
            properties:
              SKU:
                type: string
                example: PROD001
              Name:
                type: string
                example: Example Product
    BulkExportJobStatus:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-1234-567890abcdef"
        status:
          type: string
          enum: [QUEUED, PROCESSING, COMPLETED, FAILED]
          example: QUEUED
        message:
          type: string
          example: Bulk export job queued for processing.
        downloadUrl:
          type: string
          format: url
          example: http://localhost:8080/downloads/export_a1b2c3d4.csv
...